Add field defaults and kw_only flags. Fix #11
Add default fields in addition to and in lieu of Optional types,
allowing for less verbose constructor calls at the cost of losing
positional arguments due to the need for kw_only fields.

This agrees with the convention of demanding an explicit null argument
only when None is actually meaningful. Using only keyword fields relaxes
signature restrictions on using dataclasses with inheritance; in
particular, the default pre-inference pipeline can be set to be the
identity, closing #11.