Changed behaviour of DataSpec to allow for unspec'd repo. Close #18
Allow optional spec of data repo in DataSpec, which requires changes in
behaviour to getters and has'ers. Change behaviour of expand_paths to
return relative paths to on-disk data if data_repo isn't spec'd.

get_* return None exactly when their respective attributes haven't been
spec'd, and throw exceptions when the specs exist but are invalid. The
meaning of paths returned by expand_paths depends on the spec state.