Add default commit to DataSpec when repo is spec'd
Add post init which initializes DataSpec commit to HEAD if the commit is
unspec'd but the repo is, and throws an error if the commit is spec'd
but the repo isn't.

This finishes implementing the behaviour of DataSpec as follows: If the
user wants data from a particular revision, spec a repo and commit. If
the user wants the most recent revision's data, spec a repo and "HEAD"
or no commit. If the user wants uncontrolled data or data from
a worktree, specify neither.